refactor(l4-opcpa): address Copilot review comments
- schema: trim PV names + labels (.trim().min(1)) so whitespace-only strings
  are rejected at build, not silently subscribed at runtime
- gen:schema: split pure buildSchema into scripts/build-laser-schema.ts; the
  writer (gen-laser-schema.ts) runs unconditionally — drops the fragile
  process.argv[1] entry-detection guard (drift test imports the pure builder)
- ChillersSection: use c.flow (validated-unique PV) as React key, not c.label
- tests: cover whitespace-only PV rejection
